+++ app/global.d.ts
@@ -1,4 +1,4 @@
-import type { Database as DB } from "@/lib/database.types.ts";
+import type { Database as DB } from "@/lib/database.types";
 
 declare global {
   type Database = DB;
