+++ app/lockin/_components/CompletedTasks.tsx
@@ -17,15 +17,9 @@ const CompletedTasks = () => {
 
   const handleMarkTaskIncomplete = async (task: Task) => {
     const updatedTask = await markTaskIncomplete(task);
-    const taskExists = toDos!.some(todo => todo.task_id === updatedTask.task_id);
-
-    const updatedToDos = taskExists
-      ? toDos!.map(todo =>
-        todo.task_id === updatedTask.task_id ? updatedTask : todo
-      )
-      : [...toDos!, updatedTask];
-
-    console.log(updatedToDos);
+    const updatedToDos = toDos!.map(todo =>
+      todo.task_id === updatedTask.task_id ? updatedTask : todo
+    );
     setToDos(updatedToDos);
     const updatedCompleted = completedTasks!.filter((t) => t.task_id !== updatedTask.task_id);
     setCompletedTasks(updatedCompleted);
