+++ package.json
@@ -40,7 +40,7 @@
     "@types/node": "22.9.0",
     "@types/react": "^18.3.12",
     "@types/react-dom": "18.3.1",
-    "supabase": "^2.20.5",
+    "supabase": "^2.15.8",
     "tailwind-merge": "^2.6.0",
     "tailwindcss-animate": "^1.0.7",
     "typescript": "5.6.3"
